requirements.txt
numpy==1.18.5
pandas==1.0.5
Pillow==8.4.0
streamlit==1.2.0
streamlit_lottie==0.0.2
